requirements.txt
pyaml>=14.05.7
